+++ index.html
@@ -5,8 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Finance Dashboard</title>
   </head>
-  <body>
-    <div id="root"></div>
+  <body class="h-lvh flex justify-center items-center">
+    <div id="root" class="w-[750px] h-[900px] rounded-[35px] bg-[#4a4e69] shadow-lg"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
+
