+++ src/StockListsEmpty.jsx
@@ -1,0 +1,10 @@
+
+function StockListsEmpty() {
+  return (
+  <div>
+    <p>No stocks added yet.</p>
+  </div>
+  )
+}
+
+export default StockListsEmpty
