+++ src/index.css
@@ -3,7 +3,7 @@
 @tailwind components;
 @tailwind utilities;
 body {
-  background-color: #f3fbff;
+  background-color: #22223b;
 }
 input {
   border: 1px solid;
